# Preallocate string size for performance