@dom96 2021-10-30
Nice! I've been doing some benchmarking and it looks like the mark and sweep GC performs significantly better. I was able to get current HEAD to 1.5 million QPS. With this PR this goes up to 1.65 million QPS! This is definitely a hot path.

I think there is even more room here for improvements, we can get rid of allocating this string completely. I'm currently playing with this: https://gist.github.com/dom96/a041fabecd346579744c3b78ba599ec9. With the following results:

```
name ............................... min time      avg time    std dv   runs
% formatting ...................... 39.835 ms     40.321 ms    ±0.164   x124
concating ......................... 15.964 ms     16.314 ms    ±0.139   x307
smart concating ................... 11.044 ms     11.285 ms    ±0.069   x443
pre-alloc concating ................ 5.884 ms      5.983 ms    ±0.043   x835
```

This will require some special casing in httpbeast for small responses, but I'm excited to see how much faster it will make it.