@ajusa 2021-10-31
Yeah, the tricky bit is finding the things on the hot path - after that, we can just extract out the bit of code that we need, and use benchy to write a faster version. The only exception to this are the OS/system calls. We don't have as much insight into those (such as the performance metrics for them).

If there are other hot paths anyone is able to find, opening an issue would be a good first step to getting others to optimize the code!